Basic WIP ansible based config
This is is an initial release of my dotfiles and ansible based config.
It's a partial conversion of a set of scripts I've been using from a private
repo now released in a public dotfiles repo to give back to others.

Right now I'm going through a fresh install of Fedora 40 and checking
the scripts during the conversion. Further updates over the next few
days should provide a more complete config.